Add Realm crate and launch test for arm CCA
Add the Realm crate that contains wrappers for building a Realm VM.
Also, add the launch test for arm CCA that shows how to use this crate.
The test is based on an MMIO exit to figure out that the guest has
successfully executed. Note that the hlt instruction does not trigger a
VM_EXIT in cca guests since it is captured by the RMM. To trigger the
MMIO, guest has to access to an addr outside the MAX_IPA.

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>